establish proof of concept for game objects
game objects have standard values, such as an x and y value, attributes,
which are specific to the type of object and represent the bulk of it's state,
and behavior, which is a function that takes a map and returns a map,
the first map being the current state of the game object and the returned
map being the future state of the same game object.

the behavior function can store state in the attributes between moments,
including functions, since it has total control over the internal state of
game object.

this commit only introduces behavior for fireballs, and that behavior is as
yet incomplete, breaking on impact remains to be implimented