add add-millis-per-frame setting
min-millis-per-frame is the minimum length of a frame.
If it's set to 10 millis, and it naturally takes 5 millis to calculate a
frame, than the process will sleep for 5 millis.

The function of add-millis-per-frame is to guarantee some downtime
for the processor. So if it is set to 5, and a frame finishes in 15
millis, the cpu will sleep for an additional 5 millis. The same amount
of additional sleep would be applied if a frame took 1000 millis.