MVP authentication to prevent public access to some api endpoints
the default password hash will match an empty string, so leave it in
place if you want to authenticate by submitting a blank auth form.
If you want an actual password, run (hash "my password") in your repl to
generate a hash and put the hash in config.edn.

Please for the love of all that is sacred do not use a password that you
actually want to keep secret, this is not intended to be even the
slightest bit secure, it just keeps people from accidently killing the live
server and provides a solid middleware/session pattern for compojure
routes. The hashing algorithm is not intended to be secure and the hash
of your password is accesible via the /api/settings endpoint.