refactor the idea of time
add pause, unpause, and justify-time functions.
Pause ceaces the dispatching of new updates to the game-pieces and the
sending of client updates.
Unpause resumes this process.

Since updates are based on the actual flow of time and not a lockstep
with the game's framerate, unpausing will cause the game to attempt to
catch up with the real amount of time that has passed. Justify-time will
catch time up without effecting the in game world, creating the
impression that unpausing picks up right where pausing left off.